+++ app/src/interfaces/teams/team.ts
@@ -1,3 +1,5 @@
+import { IconName } from 'lucide-react/dynamic';
+
 import { Club } from '@/teamized/interfaces/club/club';
 import { ID } from '@/teamized/interfaces/common';
 
@@ -8,7 +10,7 @@ export interface Team {
     name: string;
     description: string;
     color: string;
-    icon: string;
+    icon: IconName;
     club: Club | null;
     membercount: number;
     member?: Member;
@@ -18,5 +20,5 @@ export interface TeamRequestDTO {
     name: string;
     description: string;
     color: string;
-    icon: string;
+    icon: IconName;
 }